+++ public/knockout/kbLight.js
@@ -1,39 +1,20 @@
-var ipList = [];
-/*
-function drawRow(rowData, scope) {
-    var row = $("<tr />")
-    $("#devices_table").append(row); //this will append tr element to table... keep its reference for a while since we will add cels into it
-
-	row.append($("<th scope=\"row\">"+ scope +"</th>"))
-    row.append($("<td>" + rowData.ip + "</td>"));
-    row.append($("<td>" + rowData.name + "</td>"));
-    row.append($("<td>" + rowData.description + "</td>"));
+function drawData(server){
+	var group = null;
+	
+	/* location */
+	for (var i=0; i < server.length; i++) {
+		group = $("<optgroup label=\""+ server[i].location+"\"/>");
+		$("#device_selectBox").append(group);
+		for (var j=0; j< server[i].device.length; j++) {  
+			//group.append($("<option>"+ server[i].device[j].name +"</option>"));
+			group.append($("<option value=\""+ server[i].device[j].ip +"\">"+ server[i].device[j].name +"</option>"));
+		}
+		group.append($("</optgroup>"))
+	}
 }
-
-function drawTable(data) {
-    for (var i = 0; i < data.length; i++) {
-        drawRow(data[i],i);
-    }
-}
-*/
-function drawOption(optData, scope) {
-	  var option = $("<option />");
-	  $("#device_selectBox").append(option);
-	  
-	  option.append($("<option>"+ optData.name +"</option>"))
-}
-
-function drawData(data) {
-    for (var i = 0; i < data.length; i++) {
-        drawOption(data[i],i);
-    }
-}
-
-
 
 var ViewModel = function(ip) {
     
-	/* read json */
 	$.getJSON("/settings/config", function(data) { 
 		var config = data;
     
@@ -46,5 +27,5 @@ var ViewModel = function(ip) {
 	//return this.config;
 	return;
 };
-
 ko.applyBindings(new ViewModel("127.0.0.1")); // This makes Knockout get to work
+