+++ RELEASE_NOTES.md
@@ -1,7 +1,7 @@
 version 0.0.2
 ============
 	
-  - add hue slider
+  - add hue/brightness slider
   - set several devices in 'config.json'
 
 version 0.0.1